# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_cocursor_backend_internal_domain_cursor.CodeBlock:
    properties:
      code:
        description: 代码内容
        type: string
      language:
        description: 语言（如 "go", "typescript"）
        type: string
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.ComposerData:
    properties:
      composerId:
        description: 会话唯一标识符
        type: string
      contextUsagePercent:
        description: 代码变更统计（关键字段）
        type: number
      createdAt:
        description: 创建时间戳（毫秒）
        type: integer
      createdOnBranch:
        description: 创建时的 Git 分支
        type: string
      filesChangedCount:
        description: 变更的文件数量
        type: integer
      forceMode:
        description: 强制模式
        type: string
      hasBlockingPendingActions:
        type: boolean
      hasUnreadMessages:
        description: 状态标志
        type: boolean
      isArchived:
        type: boolean
      isDraft:
        type: boolean
      isSpec:
        type: boolean
      isWorktree:
        type: boolean
      lastUpdatedAt:
        description: 最后更新时间戳（毫秒）
        type: integer
      name:
        description: 会话名称
        type: string
      numSubComposers:
        description: 层级关系
        type: integer
      referencedPlans:
        description: 引用的计划列表
        items:
          type: string
        type: array
      subtitle:
        description: 涉及的文件列表（逗号分隔）
        type: string
      totalLinesAdded:
        description: 总添加行数
        type: integer
      totalLinesRemoved:
        description: 总删除行数
        type: integer
      type:
        description: 基础信息
        type: string
      unifiedMode:
        description: 模式信息
        type: string
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.DailyCodeChanges:
    properties:
      date:
        description: 日期 YYYY-MM-DD
        type: string
      files_changed:
        description: 变更文件数
        type: integer
      lines_added:
        description: 添加行数
        type: integer
      lines_removed:
        description: 删除行数
        type: integer
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.EfficiencyMetrics:
    properties:
      avg_context_usage:
        description: 平均上下文使用率
        type: number
      avg_session_entropy:
        description: 平均会话熵值
        type: number
      entropy_trend:
        description: 熵值趋势
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.EntropyTrendItem'
        type: array
    type: object
  github_com_cocursor_backend_internal_domain_cursor.EntropyTrendItem:
    properties:
      date:
        description: 日期 YYYY-MM-DD
        type: string
      value:
        description: 熵值
        type: number
    type: object
  github_com_cocursor_backend_internal_domain_cursor.FileReference:
    properties:
      file_name:
        description: 文件名
        type: string
      file_type:
        description: 文件类型（从扩展名推断）
        type: string
      reference_count:
        description: 引用次数
        type: integer
    type: object
  github_com_cocursor_backend_internal_domain_cursor.Message:
    properties:
      code_blocks:
        description: 代码块（如果有）
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.CodeBlock'
        type: array
      files:
        description: 引用的文件（如果有）
        items:
          type: string
        type: array
      text:
        description: 消息文本
        type: string
      timestamp:
        description: 时间戳（毫秒）
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.MessageType'
        description: 消息类型：user/ai
    type: object
  github_com_cocursor_backend_internal_domain_cursor.MessageType:
    enum:
    - user
    - ai
    type: string
    x-enum-comments:
      MessageTypeAI: AI 消息
      MessageTypeUser: 用户消息
    x-enum-descriptions:
    - 用户消息
    - AI 消息
    x-enum-varnames:
    - MessageTypeUser
    - MessageTypeAI
...
  github_com_cocursor_backend_internal_domain_cursor.SessionDetail:
    properties:
      has_more:
        description: 是否还有更多消息
        type: boolean
      messages:
        description: 消息列表（按时间排序）
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.Message'
        type: array
      session:
        allOf:
        - $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.ComposerData'
        description: 会话元数据
      total_messages:
        description: 总消息数
        type: integer
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.TimeDistributionItem:
    properties:
      count:
        description: 该时段的活跃次数
        type: integer
      day:
        description: 星期几（0=周日, 1=周一, ...）
        type: integer
      hour:
        description: 小时（0-23）
        type: integer
    type: object
  github_com_cocursor_backend_internal_domain_cursor.TokenByType:
    properties:
      chat:
        description: 普通聊天
        type: integer
      composer:
        description: Composer 模式
        type: integer
      tab:
        description: Tab 自动补全
        type: integer
    type: object
  github_com_cocursor_backend_internal_domain_cursor.TokenUsage:
    properties:
      by_type:
        allOf:
        - $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.TokenByType'
        description: 按类型分类
      date:
        description: 日期 YYYY-MM-DD
        type: string
      total_tokens:
        description: 总 Token 数
        type: integer
      trend:
        description: 趋势（与昨日对比），如 "+15%" 或 "-5%"
        type: string
    type: object
  github_com_cocursor_backend_internal_domain_cursor.WorkAnalysis:
    properties:
      code_changes_trend:
        description: 代码变更趋势
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.DailyCodeChanges'
        type: array
      efficiency_metrics:
        allOf:
        - $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.EfficiencyMetrics'
        description: 效率指标
      overview:
        allOf:
        - $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkAnalysisOverview'
        description: 概览指标
      time_distribution:
        description: 时间分布（用于热力图）
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.TimeDistributionItem'
        type: array
      top_files:
        description: Top N 文件引用
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.FileReference'
        type: array
    type: object
  github_com_cocursor_backend_internal_domain_cursor.WorkAnalysisOverview:
    properties:
      acceptance_rate:
        description: 平均接受率
        type: number
      active_sessions:
        description: 活跃会话数
        type: integer
      files_changed:
        description: 变更文件数
        type: integer
      total_lines_added:
        description: 总添加行数
        type: integer
      total_lines_removed:
        description: 总删除行数
        type: integer
    type: object
...
  github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
  github_com_cocursor_backend_internal_interfaces_http_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:19960
info:
  contact: {}
  description: cocursor 守护进程 API 服务
  title: cocursor Daemon API
  version: "1.0"
paths:
...
  /sessions/{sessionId}/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 会话 ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: 限制消息数量，默认 100，最大 1000
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.SessionDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 获取会话详情
      tags:
      - 会话
  /sessions/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目名称
        in: query
        name: project_name
        type: string
      - description: 每页条数，默认 20，最大 100
        in: query
        name: limit
        type: integer
      - description: 偏移量，默认 0
        in: query
        name: offset
        type: integer
      - description: 搜索关键词（会话名称）
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.ComposerData'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 获取会话列表
      tags:
      - 会话
...
  /stats/token-usage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 日期，格式 YYYY-MM-DD，默认今天
        in: query
        name: date
        type: string
      - description: 项目名称，如 cocursor（可通过 GET /api/v1/project/list 查看所有项目）
        in: query
        name: project_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.TokenUsage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 获取 Token 使用统计
      tags:
      - 统计
  /stats/work-analysis:
    get:
      consumes:
      - application/json
      parameters:
      - description: 开始日期，格式 YYYY-MM-DD，默认最近 7 天
        in: query
        name: start_date
        type: string
      - description: 结束日期，格式 YYYY-MM-DD，默认今天
        in: query
        name: end_date
        type: string
      - description: 项目名称，如 cocursor（不提供则跨项目聚合）
        in: query
        name: project_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkAnalysis'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 获取工作分析数据
      tags:
      - 统计
schemes:
- http
swagger: "2.0"